+++ math_operations/Operations.java
@@ -1,26 +1,21 @@
 public class Operations {
-
     public int sum(int first, int second) {
         return first + second;
     }
-
-    public int substraction(int first, int second){
-        return first-second;
-
-    }
+    public int substraction(int first, int second){ return first - second; }
     public int multiply(int first, int second){
-        return first*second;
+        return first * second;
     }
     public int modulus(int first, int second){
-        return first%second;
+        return first % second;
     }
-    public float division(float first, float second){
-        return first/second;
+    public double division(double first, double second){
+        return first / second;
     }
     public int increment (int first){
         return ++first;
     }
     public boolean equality (long first, long second){
-        return first==second;
+        return first == second;
     }
 }
